button share

+++ app/src/main/java/com/example/manav/tindermaths/QuizUI.java
@@ -80,7 +80,7 @@ public class QuizUI extends AppCompatActivity implements OnGestureListener {
                 }
                 long endTime = System.currentTimeMillis();
                 long timeTaken = endTime - startTime;
-                float timeTakenSeconds = timeTaken / 1000;
+                double timeTakenSeconds = timeTaken / 1000.0;
                 Intent i = new Intent(QuizUI.this, EndGame.class);
                 i.putExtra("score", "You got " + (score*100)/20 + "% accuracy");
                 i.putExtra("time", "It took you " + timeTakenSeconds + " seconds to complete 20 questions");
